Divided the project into Modules
Created the modules ui, model, engine and passed each class the right
module.
In addition, added Jar artifacts to be created when code is compiled for
each module.
