+++ Transpool/.idea/artifacts/engine_jar.xml
@@ -1,0 +1,10 @@
+<component name="ArtifactManager">
+  <artifact name="engine:jar">
+    <output-path>$PROJECT_DIR$/out/artifacts/engine_jar</output-path>
+    <root id="root">
+      <element id="archive" name="engine.jar">
+        <element id="module-output" name="engine" />
+      </element>
+    </root>
+  </artifact>
+</component>
