+++ Transpool/engine/src/engine/Engine.java
@@ -1,0 +1,11 @@
+package engine;
+
+import model.Enums.UserAction;
+import model.Interfaces.IEngine;
+
+public class Engine implements IEngine {
+    @Override
+    public String commitAction(UserAction action) {
+        return null;
+    }
+}
