+++ Transpool/engine/src/engine/TripsManager.java
@@ -1,0 +1,22 @@
+package engine;
+
+import model.TripOffer;
+import model.TripRequest;
+
+import java.util.Set;
+
+/**
+ * engine.TripsManager is a Util class, that managers the offers and requests of the trips
+ */
+public class TripsManager {
+    private static Set<TripOffer> offers;
+    private static Set<TripRequest> requests;
+
+    public static boolean addOffer() {
+        return false;
+    }
+
+    public static boolean addRequest() {
+        return false;
+    }
+}
