+++ Transpool/model/src/model/CustomExceptions/RoadDoesNotExistException.java
@@ -1,0 +1,4 @@
+package model.CustomExceptions;
+
+public class RoadDoesNotExistException extends Exception {
+}
