+++ Transpool/model/src/model/Enums/RepetitionRate.java
@@ -1,0 +1,9 @@
+package model.Enums;
+
+public enum RepetitionRate {
+    ONE_TIME,
+    DAILY,
+    BI_DAILY,
+    WEEKLY,
+    MONTHLY
+}
