+++ Transpool/model/src/model/TripRequest.java
@@ -1,0 +1,14 @@
+package model;
+
+import java.time.LocalTime;
+
+public class TripRequest {
+    private String requestingUserName;
+    private String wantedSourceStationName;
+    private String wantedDestStationName;
+    private int wantedTripStartDay;
+    private LocalTime wantedTripStartTime;
+    private int wantedTripEndDay;
+    private LocalTime wantedTripEndTime;
+    boolean doesUserAgreeToStationExchange;
+}
