+++ Transpool/model/src/model/TripTiming.java
@@ -1,0 +1,17 @@
+package model;
+
+import model.Enums.RepetitionRate;
+
+import java.time.LocalTime;
+
+public class TripTiming {
+    private int day;
+    private LocalTime time;
+    private RepetitionRate repetitionRate;
+
+    public TripTiming(int day, LocalTime time, RepetitionRate repetitionRate) {
+        this.day = day;
+        this.time = time;
+        this.repetitionRate = repetitionRate;
+    }
+}
