--- Transpool/src/Engine/Engine.java
@@ -1,11 +1,0 @@
-package Engine;
-
-import model.Enums.UserAction;
-import model.Interfaces.IEngine;
-
-public class Engine implements IEngine {
-    @Override
-    public String commitAction(UserAction action) {
-        return null;
-    }
-}
