--- Transpool/src/Engine/TripsManager.java
@@ -1,22 +1,0 @@
-package Engine;
-
-import model.TripOffer;
-import model.TripRequest;
-
-import java.util.Set;
-
-/**
- * TripsManager is a Util class, that managers the offers and requests of the trips
- */
-public class TripsManager {
-    private static Set<TripOffer> offers;
-    private static Set<TripRequest> requests;
-
-    public static boolean addOffer() {
-        return false;
-    }
-
-    public static boolean addRequest() {
-        return false;
-    }
-}
