--- Transpool/src/Model/CustomExceptions/RoadDoesNotExistException.java
@@ -1,4 +1,0 @@
-package Model.CustomExceptions;
-
-public class RoadDoesNotExistException extends Exception {
-}
