--- Transpool/src/Model/Enums/RepetitionRate.java
@@ -1,9 +1,0 @@
-package Model.Enums;
-
-public enum RepetitionRate {
-    ONE_TIME,
-    DAILY,
-    BI_DAILY,
-    WEEKLY,
-    MONTHLY
-}
