+++ Transpool/ui/src/META-INF/MANIFEST.MF
@@ -1,0 +1,4 @@
+Manifest-Version: 1.0
+Main-Class: ui.Main
+Class-Path: lib/engine.jar lib/model.jar
+
