+++ Transpool/ui/src/ui/Main.java
@@ -1,0 +1,14 @@
+package ui;
+
+public class Main {
+
+    public static void main(String[] args) {
+//        try {
+//            model.Map.init(5, 5, new HashSet<>(), new HashSet<>());
+//        } catch (OperationNotSupportedException e) {
+//            e.printStackTrace();
+//        }
+        TranspoolConsole.start();
+
+    }
+}
